requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.2.1
Django==2.0.3
django-rest-swagger==2.1.2
djangorestframework==3.7.7
graphviz==0.8.2
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
junos-eznc==2.1.7
jxmlease==1.0.1
lxml==4.2.1
MarkupSafe==1.0
mysqlclient==1.3.12
ncclient==0.5.3
netaddr==0.7.19
openapi-codec==1.3.2
paramiko==2.4.1
pip==8.1.1
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pyserial==3.4
pytz==2018.3
PyYAML==3.12
requests==2.18.4
scp==0.10.2
setuptools==39.0.1
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
wheel==0.30.0
